react-workers: factor-out common code, improve debugging DX
Refactor: move functional core of react-workers to generic (testable/importable) moduele.

Improve DX `worker-loader` produces incorrect source-maps for `features.worker.js` so
minimizing it allows to _see_ and debug code of "business" functions.
